Split install module into smaller ones
